+++ src/config/interface.ts
@@ -9,6 +9,7 @@
 import type { LogConfig } from 'core/log';
 import type { PerfConfig } from 'core/perf/config';
 import type { OnlineCheckConfig } from 'core/net';
+import type { RequestConfig } from 'core/request';
 import type { KVStorageConfig } from 'core/kv-storage/engines/node-localstorage';
 
 export interface Config {
@@ -32,6 +33,11 @@ export interface Config {
 	 * Options for the "core/kv-storage" module
 	 */
 	kvStorage: KVStorageConfig;
+
+	/**
+	 * Options for the "core/request" module
+	 */
+	request: RequestConfig;
 
 	/**
 	 * Options for the "core/log" module