+++ src/core/request/engines/node/index.ts
@@ -13,6 +13,8 @@
 
 import got, { Options, Response as GotResponse } from 'got';
 
+import config from 'config';
+
 import AbortablePromise from 'core/promise/abortable';
 import { isOnline } from 'core/net';
 
@@ -58,7 +60,7 @@ const request: RequestEngine = (params) => {
 		const
 			{status} = await AbortablePromise.resolve(isOnline(), p.parent);
 
-		if (!status) {
+		if (!status && !config.request.workOffline) {
 			return reject(new RequestError(RequestError.Offline));
 		}