+++ src/core/request/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.? (2023-??-??)
+
+#### :rocket: New Feature
+
+* Added the ability to enable sending requests without the Internet connection
+
 ## v3.93.1 (2023-03-14)
 
 #### :bug: Bug Fix